added setup configurations and module expansion
