+++ .gitignore
@@ -1,4 +1,3 @@
 /__pycache__
 /.vscode
-/.vs
-/CPP/.x64
+/venv
