+++ requirements.txt
@@ -1,1 +1,3 @@
-requests==2.26.0
+numpy==1.22.3
+pandas==1.4.1
+scipy==1.8.0
