+++ setup.cfg
@@ -1,23 +1,33 @@
 [metadata]
-name = brainx
-description = Neural simulation module
-author = Mario González
-license = MIT
-license_file = LICENSE
-platforms = unix, linux, osx, cygwin, win32
-classifiers =
+name = biosig
+version = 0.0.1
+author = Rionzagal (Mario González)
+author_email = mario.gzz.gal@gmail.com
+license = GNU
+license_content_type = text
+description = A biological signal simulation and evaluation package
+long_description = file: README.md, LICENSE
+long_description_content_type = text/markdown
+url = https://github.com/Rionzagal/BioSig-evaluation
+project_urls = 
+    Bug Tracker = https://github.com/Rionzagal/BioSig-evaluation/issues
+keywords = python, biology, medical signal, biological signal
+classifiers = 
+    Development Status :: 1 - Planning
+    Intended Audience :: Education, Science/Research
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
+    Operating System :: OS Independent
 
-[options]
-packages = 
-    brainx
+[options] 
+package_dir = 
+    = src
+packages = find:
 install_requires = 
-    requests>=2
-python_requires = >= 3.8
-package_dir =
-    =src
+    numpy>=1.20
+    pandas>=1.4
+    scipy>=1
+python_requires = >=3.8
 zip_safe = no
+
+[options.packages.find]
+where = src
