--- src/n_values.csv
@@ -1,21 +1,0 @@
-full name,a,b,c,d
-Tonic_Spiking,0.02,0.2,-65,6
-Phasic_Spiking,0.02,0.25,-65,6
-Tonic_Bursting,0.02,0.2,-50,2
-Phasic_Bursting,0.02,0.2,-55,0.05
-Mixed_Mode,0.02,0.2,-55,4
-Spike_Frequency Adaptation,0.01,0.2,-65,8
-Class_One_Excitability,0.02,-0.1,-55,6
-Class_Two_Excitability,0.2,0.26,-65,0
-Spike_Latency,0.02,0.2,-65,6
-Subthreshold_Oscilation,0.05,0.26,-60,0
-Resonator,0.1,0.26,-60,-1
-Integrator,0.02,-0.1,-55,6
-Rebound_Spike,0.03,0.25,-60,4
-Rebound_Burst,0.03,0.25,-52,0
-Threshold_Variability,0.03,0.25,-60,5
-Bistability,0.1,0.26,-60,0
-DAP,1,0.2,-60,-21
-Accomodation,0.02,1,-55,4
-Inhibition_Induced_Spiking,-0.02,-1,-60,8
-Inhibition_Induced_Bursting,-0.026,-1,-45,-2
